# plugins/orchestrator-backend/workflows/specs/actions-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Workflow Actions for BS API",
    "description": "Workflow Actions BS API",
    "version": "0.0.1"
  },
  "servers": [{ "url": "http://host.docker.internal:7007/api/swf" }],
  "paths": {
...
    "/actions/github:webhook": {
      "post": {
        "operationId": "github:webhook",
        "description": "Creates webhook for a repository on GitHub.",
        "requestBody": {
          "description": "Input parameters for the action github:webhook in BS",
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/github_webhook" }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Action github:webhook response",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "github_webhook": {
        "type": "object",
        "required": ["repoUrl", "webhookUrl"],
        "properties": {
          "repoUrl": {
            "title": "Repository Location",
            "description": "Accepts the format 'github.com?repo=reponame&owner=owner' where 'reponame' is the new repository name and 'owner' is an organization or username",
            "type": "string"
          },
          "webhookUrl": {
            "title": "Webhook URL",
            "description": "The URL to which the payloads will be delivered",
            "type": "string"
          },
          "webhookSecret": {
            "title": "Webhook Secret",
            "description": "Webhook secret value. The default can be provided internally in action creation",
            "type": "string"
          },
          "events": {
            "title": "Triggering Events",
            "description": "Determines what events the hook is triggered for. Default: push",
            "type": "string"
          },
          "active": {
            "title": "Active",
            "type": "boolean",
            "description": "Determines if notifications are sent when the webhook is triggered. Default: true"
          },
          "contentType": {
            "title": "Content Type",
            "type": "string",
            "enum": ["form", "json"],
            "description": "The media type used to serialize the payloads. The default is 'form'"
          },
          "insecureSsl": {
            "title": "Insecure SSL",
            "type": "boolean",
            "description": "Determines whether the SSL certificate of the host for url will be verified when delivering payloads. Default 'false'"
          },
          "token": {
            "title": "Authentication Token",
            "type": "string",
            "description": "The GITHUB_TOKEN to use for authorization to GitHub"
          }
        }
      },
...
    }
  }
}